// src/lib/types/query-keys.ts
export enum queryKey {
  USERS = "users",
  LOCATION = "locations",
  CAREGIVER = "caregiver",
  USERS_META = "users_meta",
  DASHBOARD = "dashboard",
  INSURANCE_PROVIDER = "insurance-provider",
  CONSUMERS = "consumers",
  WORK_LOG = "worked-hours",
  CONSUMER = "consumer",
  ASSIGNMENT = "assignment",
  AUTHORIZED_HOURS = "authorized-hours",
  CONSUMER_DOCUMENT = "consumer-document",
  LOCATION_REPORT = "location-report",
  INSURANCE_REPORT = "insurance-report",
  CONSUMER_REPORT = "consumer-report",
  CAREGIVER_REPORT = "caregiver-report",
}
export type ActionType = "patch" | "post";