// src/enum/enum.ts
export enum Role {
  CITYADMIN = "cityadmin",
  USER = "user",
  SUPERADMIN = "superadmin",
}

export enum Genre {
  FICTION = "Fiction",
  NONFICTION = "Nonfiction",
  SCIENCE = "Science",
  HISTORY = "History",
  FANTASY = "Fantasy",
  BIOGRAPHY = "Biography",
  OTHER = "Other",
}

export enum Status {
  ACTIVE = "active",
  INACTIVE = "inactive",
  // Assignmentstatus = "Assignmentstatus",
}

export enum AssignmentStatus {
  PENDING = "pending",
  ACTIVE = "active",
  INACTIVE = "inactive",
}